requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-environ==0.10.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0